moved consts to const file
