+++ src/App.js
@@ -5,7 +5,7 @@ import { Provider, useDispatch } from 'react-redux'
 import { compose } from 'redux'
 import store from './redux/redux-store'
 import { getProfileData } from './redux/authReducer'
-import { getAccessToken } from './api/api'
+import { getAccessToken } from './const/const'
 import Main from './pages/Main'
 import Cabinet from './pages/Cabinet'
 import Login from './pages/Login'
