+++ src/api/api.js
@@ -1,18 +1,4 @@
-import axios from 'axios'
-/* eslint-disable camelcase */
-
-export const getAccessToken = () => {
-  const token = localStorage.getItem('token')
-  return token
-}
-
-const instance = axios.create({
-  baseURL: 'http://localhost:3010/api/v1',
-  headers: {
-    Accept: 'application/json',
-    Authorization: getAccessToken()
-  }
-})
+import { instance } from '../const/const'
 
 const Api = {
   getProfile() {
